requirements.txt.txt
streamlit
google-generativeai
pillow
duckduckgo-search